fix : hydros21 로직기능 보안

+++ product/C_H_PROJECT/SensorNode/components/bsp_component/sdi12/sdi12_task.h
@@ -97,7 +97,7 @@ typedef struct {
 
 typedef struct {
   char address;
-  int depth;
+  float depth;
   float temperature;
   float ec;
 } hydros21_data_t;
